requirement.txt
streamlit
python-doctr[torch]
torch
torchvision
opencv-python-headless
openai
